+++ src/Config/Config.php
@@ -67,7 +67,7 @@ class Config
         Normalize the score to be between -1 and 1 using an alpha that
         approximates the max expected value
     */
-    public static function normalize($score, $alpha = 15)
+    public static function normalize(float $score, int $alpha = 15): float
     {
         $norm_score = $score/sqrt(($score*$score) + $alpha);
         return $norm_score;